refactoring useUsers
